requirements.txt
# ML and Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Experiment Tracking
mlflow==2.9.2

# API
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# UI
streamlit==1.29.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0

# AWS
boto3==1.34.11

# Configuration
pyyaml==6.0.1

# HTTP requests (for Streamlit to call API)
requests==2.31.0
